report,auth user,datepicker

+++ database/migrations/2021_11_24_044454_create_employees_table.php
@@ -18,7 +18,7 @@ class CreateEmployeesTable extends Migration
             
             $table->string('employee_name');
 
-            $table->integer('nid_number');
+            $table->bigInteger('nid_number');
             $table->string('email')->unique();
 
             $table->integer('phone_number');
